Fix: estado de role para o botão de denúncias
